Rearranged a bunch of code, primarily the generic query methods from BSTListView into DataRepo.models.utilities.
To avoid a circular import issue, which I believe may be due to the models imported into the dataframes code, I put the import of reduceuntil into the function in DataRepo.models.utilities that uses it.  I may move the dataframe file into DataRepo.models to avoid this frequent circular import issue, because it is primarily for queries of the existing models anyway.

I also moved the reduceuntil function itself out of DataRepo.views.utils into DataRepo.utils.func_utils.

Files checked in: DataRepo/models/utilities.py DataRepo/tests/models/test_utilities.py DataRepo/tests/views/models/bst/test_list_view.py DataRepo/tests/views/test_utils.py DataRepo/views/models/bst/list_view.py DataRepo/views/utils.py DataRepo/tests/utils/test_func_utils.py DataRepo/utils/func_utils.py
