+++ DataRepo/models/utilities.py
@@ -8,7 +8,7 @@ from django.apps import apps
 from django.conf import settings
 from django.core.exceptions import ObjectDoesNotExist, ValidationError
 from django.db import ProgrammingError
-from django.db.models import Expression, F, Field, Model
+from django.db.models import Expression, F, Field, Model, QuerySet
 from django.db.models.expressions import Combinable
 from django.db.models.fields.related_descriptors import (
     ForwardManyToOneDescriptor,
@@ -915,6 +915,418 @@ def model_title_plural(model: Type[Model]) -> str:
             return f"{camel_to_title(model.__name__)}s"
     except Exception:
         return f"{camel_to_title(model.__name__)}s"
+
+
+def get_field_val_by_iteration(
+    rec: Model,
+    field_path: List[str],
+    related_limit: int = 5,
+    sort_field_path: Optional[List[str]] = None,
+    _sort_val: Optional[List[str]] = None,
+):
+    """Private recursive method that takes a record and a path and traverses the record along the path to return
+    whatever value is at the end of the path.  If it traverses through a many-related model, it returns a list of
+    values.
+
+    NOTE: If the end of the field_path is a foreign key itself, the value (or values) returned is not the key value
+    (i.e. not an integer), but a model object (or objects).
+
+    NOTE: The recursive calls are made via the supporting methods:
+    - _get_field_val_by_iteration_onerelated_helper - Handles passing through singly-related foreign keys along the
+        field_path
+    - _get_field_val_by_iteration_manyrelated_helper - Handles passing through many-related foreign keys along the
+        field_path
+
+    The way this works is, _get_field_val_by_iteration_manyrelated_helper is called at any point along the
+    field_path (possibly multiple points), where the foreign key being passed through is many-related.  Anytime a
+    foreign key (of the end-value) along the field path is 1-related to its parent, it calls
+    _get_field_val_by_iteration_onerelated_helper.  _get_field_val_by_iteration_onerelated_helper returns a 3-member
+    tuple.  As those values are being passed back through the call stack, when they pass through the many-related
+    step, those tuples are collected into a list of tuples.  The end result will either be a tuple (if there are no
+    many-related relations along the path), or a list of tuples.  Each tuple is the value itself, a sort value, and
+    a primary key.  In the case of there being no many-related component in the field_path, the second 2 values in
+    the tuple are meaningless.
+
+    Assumptions:
+        1. The sort_field_path value will be a field under the associated column's related_model_path
+    Args:
+        rec (Model): A Model object.
+        field_path (List[str]): A path from the rec object to a field/column value, that has been split by
+            dunderscores.
+        related_limit (int) [5]: Truncate/stop at this many (many-related) records.
+        sort_field_path (Optional[List[str]]): A path from the rec object to a sort field, that has been split by
+            dunderscores.  Only relevant if you know the field path to traverse through a many-related model.
+        _sort_val (Optional[List[str]]): Do not supply.  This holds the sort value if the field path is longer than
+            the sort field path.
+        NOTE: We don't need to know if sorting is forward or reverse.  We are only returning tuples containing the
+        sort value.  The sort must be done later, by the caller.
+    Exceptions:
+        ValueError when the sort field returns more than 1 value.
+    Returns:
+        (Union[List[Tuple[Any, Any, Any]]], Tuple[Any, Any, Any]): A list of 3-membered tuples or a 3-membered
+            tuple.  Each tuple contains the value, a sort value, and a unique value.
+    """
+    if len(field_path) == 0 or rec is None:
+        return None
+
+    if is_many_related_to_parent(field_path[0], type(rec)):
+        # This method handles only fields that are many-related to their immediate parent
+        return _get_field_val_by_iteration_manyrelated_helper(
+            rec,
+            field_path,
+            related_limit=related_limit,
+            sort_field_path=sort_field_path,
+            _sort_val=_sort_val,
+        )
+
+    # This method handles only fields that are singly related to their immediate parent
+    return _get_field_val_by_iteration_onerelated_helper(
+        rec,
+        field_path,
+        related_limit=related_limit,
+        sort_field_path=sort_field_path,
+        _sort_val=_sort_val,
+    )
+
+
+def _get_field_val_by_iteration_onerelated_helper(
+    rec: Model,
+    field_path: List[str],
+    related_limit: int = 5,
+    sort_field_path: Optional[List[str]] = None,
+    _sort_val: Optional[List[str]] = None,
+):
+    """Private recursive method that takes a field_path and a record (that is 1:1 related with the first element in
+    the remaining field_path) and traverses the record along the path to return whatever ORM object's field value is
+    at the end of the path.
+
+    NOTE: Recursive calls go to get_field_val_by_iteration, which calls this method or the companion method
+    (_get_field_val_by_iteration_manyrelated_helper) for many-related portions of the field_path.
+
+    Assumptions:
+        1. The related_sort_fld value will be a field under the related_model_path
+    Args:
+        rec (Model): A Model object.
+        field_path (List[str]): A path from the rec object to a field/column value, that has been split by
+            dunderscores.
+        related_limit (int) [5]: Truncate/stop at this many (many-related) records.
+        sort_field_path (Optional[List[str]]): A path from the rec object to a sort field, that has been split by
+            dunderscores.  Only relevant if you know the field path to traverse through a many-related model.
+        _sort_val (Optional[List[str]]): Do not supply.  This holds the sort value if the field path is longer than
+            the sort field path.
+    Exceptions:
+        ValueError when the sort field returns more than 1 value.
+    Returns:
+        (Tuple[Any, Any, Any]): A tuple containing the value, a sort value, and a unique value.
+    """
+    if is_many_related_to_parent(field_path[0], type(rec)):
+        raise TypeError(
+            "_get_field_val_by_iteration_onerelated_helper called with a many-related field"
+        )
+
+    val_or_rec = getattr(rec, field_path[0])
+
+    next_sort_field_path = sort_field_path[1:] if sort_field_path is not None else None
+    # If we're at the end of the field path, we need to issue a separate recursive call to get the sort value
+    if (
+        sort_field_path is not None
+        and len(sort_field_path) > 0
+        and _sort_val is None
+        and (
+            sort_field_path[0] != field_path[0]
+            or len(sort_field_path) == 1
+            or len(field_path) == 1
+        )
+    ):
+        # NOTE: Limiting to 2, because we only expect 1 and will raise ProgrammingError if multiple returned
+        sort_val, _, _ = get_field_val_by_iteration(
+            rec, sort_field_path, related_limit=2
+        )
+        if isinstance(sort_val, list):
+            raise ProgrammingError(
+                "The sort value must not be many-related with the value for the column"
+            )
+        next_sort_field_path = None
+        _sort_val = sort_val
+
+    if len(field_path) == 1:
+        uniq_val = val_or_rec
+        if isinstance(val_or_rec, Model):
+            uniq_val = val_or_rec.pk
+        # NOTE: Returning the value, a value to sort by, and a value that makes it unique per record (or field)
+        return val_or_rec, _sort_val, uniq_val
+
+    return get_field_val_by_iteration(
+        val_or_rec,
+        field_path[1:],
+        related_limit=related_limit,
+        sort_field_path=next_sort_field_path,
+        _sort_val=_sort_val,
+    )
+
+
+def _get_field_val_by_iteration_manyrelated_helper(
+    rec: Model,
+    field_path: List[str],
+    related_limit: int = 5,
+    sort_field_path: Optional[List[str]] = None,
+    _sort_val: Optional[List[str]] = None,
+):
+    """Private recursive method that takes a field_path and a record (that is many:1_or_many related with the first
+    element in the remaining field_path) and traverses the record along the path to return values found at the end
+    of the field_path.
+
+    NOTE: Recursive calls go to get_field_val_by_iteration, which calls this method or the companion method
+    (_get_field_val_by_iteration_single_helper) for singly-related portions of the field_path.
+
+    NOTE: The recursive calls to get_field_val_by_iteration come from the 2 supporting methods:
+    - _last_many_rec_iterator
+    - _recursive_many_rec_iterator
+
+    Assumptions:
+        1. The sort_field_path value starts with the field_path
+    Args:
+        rec (Model): A Model object.
+        field_path (List[str]): A path from the rec object to a field/column value, that has been split by
+            dunderscores.
+        related_limit (int) [5]: Truncate/stop at this many (many-related) records.
+        sort_field_path (Optional[List[str]]): A path from the rec object to a sort field, that has been split by
+            dunderscores.  Only relevant if you know the field path to traverse through a many-related model.
+        _sort_val (Optional[List[str]]): Do not supply.  This holds the sort value if the field path is longer than
+            the sort field path.
+    Exceptions:
+        ValueError when the sort field returns more than 1 value.
+    Returns:
+        (List[Tuple[Any, Any, Any]]): A list of tuples, the size of which corresponds to the number of unique
+            records.
+    """
+    from DataRepo.utils.func_utils import reduceuntil
+
+    if rec is None:
+        return []
+
+    if not is_many_related_to_parent([field_path[0]], type(rec)):
+        raise TypeError(
+            "_get_field_val_by_iteration_many_related_helper called without a many-related field"
+        )
+
+    mr_qs: QuerySet = getattr(rec, field_path[0])
+
+    next_sort_field_path = sort_field_path[1:] if sort_field_path is not None else []
+    # If the sort_field_path has diverged from the field_path, retrieve its value
+    if sort_field_path is not None and sort_field_path[0] != field_path[0]:
+        sort_val, _, _ = get_field_val_by_iteration(
+            rec,
+            sort_field_path,
+            # We only expect 1 value and are going to assume that the sort field was properly checked/generated to
+            # not go through another many-related relationship.  Still, we specify the limit to be safe.
+            related_limit=1,
+        )
+        if isinstance(sort_val, list):
+            raise ProgrammingError(
+                "The sort value must not be many-related with the value for the column"
+            )
+        next_sort_field_path = []
+        _sort_val = sort_val
+
+    if len(field_path) == 1:
+
+        uniq_vals = reduceuntil(
+            lambda ulst, val: ulst + [val] if val not in ulst else ulst,
+            lambda val: related_limit is not None and len(val) >= related_limit,
+            _last_many_rec_iterator(mr_qs, next_sort_field_path),
+            [],
+        )
+
+        return uniq_vals
+
+    uniq_vals = reduceuntil(
+        lambda ulst, val: ulst + [val] if val not in ulst else ulst,
+        lambda val: related_limit is not None and len(val) >= related_limit,
+        _recursive_many_rec_iterator(
+            mr_qs,
+            field_path[1:],
+            next_sort_field_path,
+            related_limit,
+            _sort_val,
+        ),
+        [],
+    )
+
+    return uniq_vals
+
+
+def _last_many_rec_iterator(
+    mr_qs: QuerySet,
+    next_sort_field_path: List[str],
+):
+    """Private iterator to help _get_field_val_by_iteration_many_related_helper.  It iterates through the queryset,
+    converting the many-related records to tuples of the record, the sort value, and the primary key.  This allows
+    the caller to stop when it reaches its goal.  This is called when we're at the end of the field_path.  I.e. the
+    end of the field_path is a foreign key to a many-related model.  It will make a recursive call if the
+    sort_field_path is deeper than the field_path.
+
+    NOTE: This lower-cases the sort value (if it is a str).
+
+    Args:
+        mr_qs: (QuerySet): A queryset of values that are many-related to self.model.
+        next_sort_field_path (Optional[List[str]]): The next sort_field_path that can be supplied directly to
+            recursive calls to get_field_val_by_iteration without slicing it.
+    Exceptions:
+        None
+    Returns:
+        (Tuple[Any, Any, Any]): The value, sort-value, and primary key of the many-related model
+    """
+    mr_rec: Model
+    for mr_rec in mr_qs.all():
+        yield (
+            # Model object is the value returned
+            mr_rec,
+            # Each rec gets its own sort value.
+            (
+                # TODO: See if this loop always causes a query.  If it does, then this iteration strategy may not be
+                # as efficient as I'd hoped and should be entirely removed, as should the _lower() method
+                _lower(get_field_val_by_iteration(mr_rec, next_sort_field_path)[0])
+                if len(next_sort_field_path) > 0
+                # Lower-case the string version of the many-related model object
+                else str(mr_rec).lower()
+            ),
+            # We don't need pk for uniqueness when including model objects, but callers expect it
+            mr_rec.pk,
+        )
+
+
+def _recursive_many_rec_iterator(
+    mr_qs: QuerySet,
+    next_field_path: List[str],
+    next_sort_field_path: List[str],
+    related_limit: int,
+    _sort_val,
+):
+    """Private iterator to help _get_field_val_by_iteration_many_related_helper.  It iterates through the queryset,
+    retrieving the values at the end of the path using recursive calls.  This allows the caller to stop when it
+    reaches its goal.  This is called when a many-related model is encountered before we're at the end of the
+    field_path.
+
+    Args:
+        mr_qs: (QuerySet): A queryset of values that are many-related to the model the queryset comes from.
+        field_path (List[str])
+        next_sort_field_path (List[str]): In order to simplify this method, instead of taking sort_field_path (which
+            needs to be checked and converted to the next path, because it can diverge or be a different length from
+            the field_path), that work must be done before calling this method.
+        related_limit (int)
+        _sort_val (Any)
+    Exceptions:
+        None
+    Returns:
+        (Tuple[Any, Any, Any]): The value, sort-value, and primary key of the many-related model
+    """
+    mr_rec: Model
+    for mr_rec in mr_qs.all():
+        val = get_field_val_by_iteration(
+            mr_rec,
+            next_field_path,
+            related_limit=related_limit,
+            sort_field_path=next_sort_field_path,
+            _sort_val=_sort_val,
+        )
+        if isinstance(val, tuple):
+            yield val
+        else:
+            # Must be a list
+            for tpl in val:
+                yield tpl
+
+
+def get_many_related_field_val_by_subquery(
+    rec: Model,
+    field_path: str,
+    related_limit: int = 5,
+    annotations: dict = {},
+    order_bys: list = [],
+    distincts: list = [],
+) -> list:
+    """
+
+    Args:
+        rec (Model): A Model object whose Model corresponds to the start of the field_path.
+        field_path (str): A dunderscore delimited field path whose starting field is a field of the rec's model.
+        related_limit (int) [5]: Limit the size of the list returned to this many field values.
+        annotations (dict) [{}]: Any annotations that need to be created.  Primarily, this is to support what is
+            provided in the order_by list.
+        order_bys (List[Union[str, OrderBy]]) [[]]: A list of OrderBy values (object or str).  This will be
+            automatically appended to with the field_path and the primary key of the last many-related model foreign
+            key in the field_path.
+        distincts (List[str]) [[]]: A list of field paths.  Must match the fields contained in the order_bys list.
+            This will be prepended with the field_path and appended with the primary key of the last many-related
+            model foreign key in the field_path.
+    Exceptions:
+        None
+    Returns:
+        vals_list (list): A unique list of values from the many-related model at the end of the field path.
+    """
+    # This is used to ensure that the last many-related model is what is made distinct (not any other one-related
+    # model that could be later in the field_path after the many-related model)
+    many_related_model_path = field_path_to_model_path(
+        rec.__class__, field_path, many_related=True
+    )
+
+    # If the field in the field_path is a foreign key (whether it is to the many-related model or another model that
+    # is one-related with the many-related model), the primary key value returned is converted into model objects.
+    related_model_path = field_path_to_model_path(rec.__class__, field_path)
+    related_model = model_path_to_model(rec.__class__, related_model_path)
+    is_fk = related_model_path == field_path
+
+    # Append the actual field from the column (which may differ from the sort's field [c.i.p. as happens in column
+    # groups])
+    order_bys.append(field_path)
+    # Append the many-related model's primary key in order to force a value for each distinct record
+    order_bys.append(f"{many_related_model_path}__pk")
+
+    # Prepend the actual field from the column (which may differ from the sort).  We put it first so we can easily
+    # extract it in the values_list.  This is necessary because Django's combination of annotate, distinct, and
+    # values_list have some quirks that prevent an intuitive usage of just flattening with the one value you want.
+    # I tried many different versions of this before figuring this out.
+    distincts.insert(0, field_path)
+    # Append the many-related model's primary key in order to force a value for each distinct record
+    distincts.append(f"{many_related_model_path}__pk")
+
+    # We re-perform (essentially) the same query that generated the table, but for one root-table record, and with
+    # all of the many-related values joined in to "split" the row, but we're only going to keep those many-related
+    # values.  We can/will get repeated values if the field is not unique in the many-related model, but they each
+    # will represent a unique many-related model record.  Furthermore, if the column is in a column group, the
+    # order-bys will by based on the same field expression, because BSTColumnGroup modifies col.sorter.
+    qs = (
+        rec.__class__.objects
+        # Filter for the current root model record
+        .filter(pk=rec.pk)
+        # The annotation is for use in the order_by, since it needs to be supplied to .distinct()
+        .annotate(**annotations)
+        # The order_by should be the annotation (e.g. the lower-cased version of the sort field) and 2 fields
+        # necessary to be able to supply to distinct and the values_list: the primary key of the many-related model
+        # (distinct) and the field_path of the field we want for the column (values_list)
+        .order_by(*order_bys)
+        # NOTE: This makes it distinct per the target many-related model record, even if there exist multiple many-
+        # related models in the field path.
+        .distinct(*distincts)
+    )
+
+    vals_list = [
+        # Return an object, like an actual queryset does, if val is a foreign key field
+        related_model.objects.get(pk=val) if is_fk else val
+        for val in list(v[0] for v in qs.values_list(*distincts)[0:related_limit])
+    ]
+
+    return vals_list
+
+
+# TODO: This should be removed and the BSTManyRelatedColumn class should have a python sort method to take its place
+def _lower(val):
+    """Intended for use in list comprehensions to lower-case the sort value, IF IT IS A STRING.
+    Otherwise it returns the unmodified value."""
+    if isinstance(val, str):
+        return val.lower()
+    return val
 
 
 # TODO: Figure out a way to move these exception classes to exceptions.py without hitting a circular import
