+++ DataRepo/tests/views/test_utils.py
@@ -1,26 +1,9 @@
 from DataRepo.models.study import Study
 from DataRepo.tests.tracebase_test_case import TracebaseTestCase
-from DataRepo.views.utils import (
-    GracefulPaginator,
-    get_cookie_dict,
-    reduceuntil,
-)
+from DataRepo.views.utils import GracefulPaginator, get_cookie_dict
 
 
 class UtilsMainTests(TracebaseTestCase):
-
-    def test_reduceuntil(self):
-        input_list = [2, 2, 2, 2, 2, 3, 4, 5, 6, 7]
-        max_unique_len = 2
-        self.assertEqual(
-            [2, 3],
-            reduceuntil(
-                lambda ulst, val: ulst + [val] if val not in ulst else ulst,
-                lambda val: len(val) >= max_unique_len,
-                input_list,
-                [],
-            ),
-        )
 
     def test_GracefulPaginator(self):
         Study.objects.create(name="1")
